* Reduced NetworkMessage size to 15kb * Added new script functions(getCreaturePosition, isItemStackable, isItemRune, isItemDoor, isItemContainer, isItemFluidContainer) * Added new npc script function getNpcCid and deprecated some old functions * Changed some long variables to uint32_t

+++ otserv/trunk/actions.cpp
@@ -433,8 +433,8 @@ bool Action::executeUse(Player* player, Item* item,
 		env->setScriptId(m_scriptId, m_scriptInterface);
 		env->setRealPos(player->getPosition());
 	
-		long cid = env->addThing(player);
-		long itemid1 = env->addThing(item);
+		uint32_t cid = env->addThing(player);
+		uint32_t itemid1 = env->addThing(item);
 	
 		lua_State* L = m_scriptInterface->getLuaState();
